Last fix

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -3,7 +3,7 @@ package com.epam.utils;
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
         if (org.apache.commons.lang3.StringUtils.isNumeric(str)) {
-            int num = Integer.parseInt(str);
+            double num = Double.parseDouble(str);
             return num > 0;
         } else {
             return false;
